Add Rust extraction
Add an extraction to Rust. It should be more or less usable but missing
a few niceties. For one, we are not caching closures which might be
expensive when things are partially applied often. The second is that we
always include PhantomData in all constructors to allow all datatypes to
have lifetime parameters. While this should not affect performance it
makes the extracted code somewhat ugly.