Optimize PrettyPrinterMonad
Instead of having one gigantic string that is continuously concatenated
to, use a list of lines and their indentation levels and put lines in
here. Then only concatenate these at the end.

This means getting the current indentation is now an O(1) operation
and concatenating new lines and character is much much cheaper.

Any newline added is now automatically indented (unless it is empty, in
which case the indentation is not printed).