Add a plugin to derive record setters using MetaCoq
This replaces the coq-record-update plugin based on Ltac that we used to
define nice record update syntax with a similarly one (on the surface)
that uses MetaCoq to define setters. The benefits are that we can
produce better terms and that we no longer have to list each member of
records when deriving. The downside is that it does not support indexed
records.